wsclient/requirements.txt
autopep8==1.5.7
certifi==2021.10.8
charset-normalizer==2.0.7
gevent==21.8.0
greenlet==1.1.2
idna==3.3
pycodestyle==2.8.0
redis==3.5.3
redis-server==6.0.9
requests==2.26.0
toml==0.10.2
urllib3==1.26.7
websocket-client==1.2.1
websockets==10.0
zope.event==4.5.0
zope.interface==5.4.0
